Update README.md
The above C project is the code for "FLAMES". FLAMES is a fun depiction of finding out the relationship between two members by their names . The FLAMES stands for friend, lover, affection, marriage, enemy, and sibling respectively. It is like a fun game used for determining one of these relationships between two members. In this code, we get two strings(names) from the user to find out the relationship between them. Then two strings are compared for the similarity of characters and the similar characters are replaced by "*" according to the code. Using for loop, count the number of characters which is not replaced by "*" for both the strings. Here we get the total number of characters not replaced by "*".Then by using the total count ,the characters in the "FLAMES" gets replaced by "*" using the value of count. Each time, one of the characters get replaced by "*".Finally ,there will be one character remaining in the word "FLAMES". This one character depicts the relationship between two strings. As the word "FLAMES"  is initialised to an array, each character in an array will have an index.  So, the index of the character which is not replaced by "*" is found out. If the index is 0,it represents friends, 1 represents lover, 2 is for affection, and so on .This is done using if elseif statements. After finding the index ,it will print the corresponding relationship between the two strings.